examples/cosmos_checkpoint_azure/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform

azure-functions
azure-functions-langgraph[cosmos]>=0.7.0,<0.8.0
langgraph>=1.0,<2.0
langchain-core>=1.0,<2.0
